车辆入场、出场，自动更新数据库

+++ src/main/java/com/example/parkingmanagesys/Service/Impl/PeriodicUpdateServiceImpl.java
@@ -15,14 +15,16 @@ public class PeriodicUpdateServiceImpl implements PeriodicUpdateService {
     @Autowired
     private BookingInformationMapper bookingInformationMapper;
 
-    @Scheduled(cron = "0/20 * * * * *")
+    @Scheduled(cron = "5 * * * * *")
     public void deleteByBookingId() {
         List<BookingInformation> bookingInformationList=bookingInformationMapper.selectAll();
         for(int i=0;i<bookingInformationList.size();i++){
-            long diff=new Date().getTime()-bookingInformationList.get(i).getBookingTime().getTime();
-            long hour=diff%(1000*24*60*60)/(1000*60*60);
-            if(hour>=2){
-                bookingInformationMapper.deleteByBookingId(bookingInformationList.get(i).getBookingId());
+            if(bookingInformationList.get(i).getArrivalTime()==null) {
+                long diff = new Date().getTime() - bookingInformationList.get(i).getBookingTime().getTime();
+                long hour = diff / 1000 / 60 / 60;
+                if (hour >= 2) {
+                    bookingInformationMapper.deleteByBookingId(bookingInformationList.get(i).getBookingId());
+                }
             }
         }
     }
